+++ src/yooai/containers/HBox.ts
@@ -1,19 +1,10 @@
 import {Container} from "pixi.js";
 import {HorizontalLayoutBehavior} from "../..";
+import {AbstractComponent} from "../core/AbstractComponent";
 import {getHeight, getWidth} from "../layout/utils";
 import {AbstractBox} from "./AbstractBox";
 
 export class HBox extends AbstractBox {
-  public get contentWidth(): number {
-    return this._contentWidth;
-  }
-
-  public get contentHeight(): number {
-    return this._contentHeight;
-  }
-  private _contentWidth: number = 0;
-  private _contentHeight: number = 0;
-
   public constructor(parent?: Container, x: number = 0, y: number = 0) {
     super(new HorizontalLayoutBehavior(), parent, x, y);
 
@@ -21,11 +12,10 @@ export class HBox extends AbstractBox {
   }
 
   protected drawLayout(): void {
-    super.drawLayout();
     this.calculateContentSize();
-
     this._componentWidth = this._contentWidth + this.marginLeft + this.marginRight;
     this._componentHeight = this._contentHeight + this.marginTop + this.marginBottom;
+    super.drawLayout();
   }
 
   private calculateContentSize() {
@@ -40,6 +30,9 @@ export class HBox extends AbstractBox {
     while (index >= 0) {
       const child = this.children[index];
       if (child instanceof Container) {
+        if (child instanceof AbstractComponent && child.isInvalid()) {
+          child.validateNow();
+        }
         const childWidth = getWidth(child);
         const childHeight = getHeight(child);
         if (!contentWidthIsSet) {