+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+# 0.0.37 - 2020-11-30
+- Fixed: Layouts behavior
+- Feature: Added `resizeToContent` method for `Label`
+
 # 0.0.36 - 2020-11-20
 - Fixed: Added support for legacy canvas renderer